Deprecate the raw entry API in favor of `HashTable`
It is now moved under the "raw-entry" Cargo feature.

For now the raw-entry feature is kept enabled by default because
`HashSet` uses it internally, but this will be disabled by default once
this is no longer the case.
