+++ README.md
@@ -106,6 +106,7 @@ This crate has the following Cargo features:
 - `rayon`: Enables rayon parallel iterator support.
 - `equivalent`: Allows comparisons to be customized with the `Equivalent` trait.
 - `raw`: Enables access to the experimental and unsafe `RawTable` API.
+- `raw-entry`: Enables access to the deprecated `RawEntry` API.
 - `inline-more`: Adds inline hints to most functions, improving run-time performance at the cost
   of compilation time. (enabled by default)
 - `default-hasher`: Compiles with ahash as default hasher. (enabled by default)
