+++ ci/run.sh
@@ -43,8 +43,8 @@ if [ "${CHANNEL}" = "nightly" ]; then
 fi
 
 # Make sure we can compile without the default hasher
-"${CARGO}" -vv build --target="${TARGET}" --no-default-features
-"${CARGO}" -vv build --target="${TARGET}" --release --no-default-features
+"${CARGO}" -vv build --target="${TARGET}" --no-default-features --features raw-entry
+"${CARGO}" -vv build --target="${TARGET}" --release --no-default-features --features raw-entry
 
 "${CARGO}" -vv ${OP} --target="${TARGET}"
 "${CARGO}" -vv ${OP} --target="${TARGET}" --features "${FEATURES}"
